+++ packages/signals/lib/src/list_signal.dart
@@ -3,7 +3,7 @@ import 'dart:math';
 import 'signals.dart';
 
 /// A [Signal] that holds a [List].
-class ListSignal<E> extends Signal<List<E>> implements List<E> {
+class ListSignal<E> extends ValueSignal<List<E>> implements List<E> {
   /// Creates a [ListSignal] with the given [value].
   ListSignal(super.value, {super.debugLabel});